feat: implement WritableStream monitoring strategy
It wraps a WritableStream, using the pipeTo() pattern used to connect
ReadableStreams to WritableStreams. The aim is to expose the `closed`
promise from the wrapped WritableStream, so that code can react to the
wrapped stream closing without holding a writer on it (which prevents
it being used in a pipeline or written to by another stream).

The monitor intentionally doesn't implement support for invoking an
error handler, nor handling any resulting errors with stream aborts.
Those details are left to callers, via the promise returned from
pipeTo(). This allows the monitor to be pretty simple.