feat: implement ShutdownAwareTransformStream
A TransformStream, except that (exactly like WritableStreams) the
controller has an AbortSignal on its `signal` which fires when the
stream is aborted. In conjunction with the flush() method, this signal
can be used to clean up resources held by a transformer.

The stream's Transformer object can define a close() method which is
always called when the stream closes or errors.