fix!: change write proxying behaviour of monitor stream
Previously it was following the advice of the streams spec and not
awaiting the promise returned from writer.write(). However this does
not result in desirable behaviour for our use case of
proxying/mimicking the monitored stream.

By not awaiting the promise, we cause our write() method not to throw
when the monitored stream's write fails. Instead we would throw on the
next write() as a result of awaiting the writer.ready promise.
Additionally, we hide the highWaterMark of the monitored stream, and
introduce 1 extra chunk of queue capacity into the queue.

We can avoid both of these issues by awaiting (returning) the monitored
stream's write() promise from our write(), as long as we use whichever
queuing strategy the monitored stream would have used on the monitor
stream. The 1 extra chunk of queue capacity is also removed with this
method, as by awaiting its write() promise, we only initiate a write()
on the monitored stream after it's queue has become emptied (but this
is not a problem so long as the monitor stream has the desired queue
capacity).